backend/requirements.txt
asgiref==3.8.1
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
Django==5.2.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
fonttools==4.54.1
joblib==1.5.2
kiwisolver==1.4.7
lxml==6.0.2
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.3.3
pdfminer.six==20251107
pillow==11.0.0
pycparser==2.23
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
reportlab==4.4.3
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
